Clean up Iced GUI experiment files - keep working version as iced_gui.rs
